Move providers to specific folder
