+++ src/classes/index.ts
@@ -1,5 +1,3 @@
 export { AwsConstruct } from "./AwsConstruct";
-export { AwsProvider } from "./AwsProvider";
-export { StripeProvider } from "./StripeProvider";
 export type { ConstructInterface, StaticConstructInterface } from "./Construct";
 export type { ProviderInterface, StaticProviderInterface } from "./Provider";
